requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-chroma
pypdf
python-docx
dashscope